requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-ollama==0.1.3
langchain-chroma==0.1.4
langchain-huggingface==0.0.3
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf==4.3.1
streamlit==1.37.1
python-dotenv==1.0.1
tqdm==4.66.5
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
greenlet==3.0.3
posthog<4
